rename: my 폴더 my-page로 이름 수정
