+++ src/components/my-page/PreviewList.tsx
