+++ src/components/my-page/previewList.module.scss
