--- src/components/sidebar/myPageSidebarTemp.module.scss
@@ -1,32 +1,0 @@
-.sidebar {
-  width: 220px;
-  height: 100%;
-  //border-right: 1px solid #ddd;
-  padding: 20px;
-  background: #fff;
-
-  .sidebarTitle {
-    font-size: 18px;
-    font-weight: bold;
-    margin-bottom: 20px;
-  }
-
-  .nav {
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-
-    .navItem {
-      cursor: pointer;
-      font-size: 14px;
-      color: #333;
-      padding: 8px;
-      border-radius: 5px;
-
-      &.active {
-        font-weight: bold;
-        color: #007aff;
-      }
-    }
-  }
-}
