--- src/components/tab-navigation/tabNavigation.module.scss
@@ -1,29 +1,0 @@
-.tabs {
-  @include flex-center;
-  //gap:20px;
-  margin-bottom: 20px;
-  border-bottom: 2px solid #ddd;
-}
-
-.tabItem {
-  flex: 1;
-  padding: 10px 20px;
-  cursor: pointer;
-  font-size: 20px;
-  text-align: center;
-  color: #999;
-  transition:
-    background-color 0.3s,
-    color 0.3s;
-  border-bottom: 2px solid transparent;
-
-  &.activeTab {
-    color: #007aff;
-    border-bottom: 3px solid #007aff;
-    font-weight: bold;
-  }
-
-  &:hover {
-    background-color: #f0f0f0;
-  }
-}
