+++ src/pages/layout/MyPageLayout.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from 'react-router-dom';
-import { Sidebar } from '../../components/my/Sidebar';
+import { Sidebar } from '../../components/my-page/Sidebar';
 import classNames from 'classnames';
 import styles from './myPageLayout.module.scss';
