+++ src/pages/my-page/myPage/RecentNews.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import classNames from 'classnames';
 import styles from './recentNews.module.scss';
 import { PreviewList } from '../../../components/my-page/PreviewList';
-import Arrow from '../../assets/icons/arrowRight.svg?react';
+import Arrow from '../../../assets/icons/arrowRight.svg?react';
 
 export const RecentNews = () => {
   const [isOpenMoreDelete, setIsOpenMoreDelete] = useState<boolean>(false);
