+++ src/pages/my-page/myPagePosts/MyPagePosts.tsx
@@ -1,10 +1,7 @@
 import { useState } from 'react';
 import styles from './myPagePosts.module.scss';
 //import classNames from 'classnames';
-import { Header } from '../../../components/common/header/Header.tsx';
-import MyPageSidebar from '../../../components/sidebar/MyPageSidebar.tsx';
 import PostCard from '../../../components/postcard/PostCard.tsx';
-import TabNavigation from '../../../components/tab-navigation/TabNavigation.tsx';
 
 const TABS = ['아이디어 마켓', '요청과제', '협업광장'];
 
@@ -59,20 +56,11 @@ export const MyPagePosts = () => {
 
   return (
     <div className={styles.container}>
-      <Header />
       <div className={styles.contentWrapper}>
-        <MyPageSidebar />
-
         <main className={styles.mainContent}>
           <h1 className={styles.title}>
             게시물 관리 <span className={styles.count}>11</span>
           </h1>
-
-          <TabNavigation
-            tabs={TABS}
-            activeTab={activeTab}
-            onTabChange={setActiveTab}
-          />
 
           {/* 게시물 리스트 */}
           <div className={styles.postList}>
